requirements.txt
APScheduler==3.6.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.1
Flask-Caching==1.10.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.8.6
pymongo==3.11.4
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
Werkzeug==2.0.1
